Add Google login page and account-based storage
